Pull test env. classes under the `io.spine.logging.given`
