+++ logging/src/jvmTest/kotlin/io/spine/logging/CountingRateLimiterSpec.kt
@@ -33,7 +33,7 @@ import io.kotest.matchers.types.shouldBeSameInstanceAs
 import io.spine.logging.LogContext.Key
 import io.spine.logging.RateLimitStatus.Companion.DISALLOW
 import io.spine.logging.backend.given.FakeMetadata
-import io.spine.logging.jvm.given.FakeLogSite
+import io.spine.logging.given.FakeLogSite
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
