+++ logging/src/jvmTest/kotlin/io/spine/logging/RateLimitStatusSpec.kt
@@ -36,7 +36,7 @@ import io.spine.logging.RateLimitStatus
 import io.spine.logging.RateLimitStatus.Companion.ALLOW
 import io.spine.logging.RateLimitStatus.Companion.DISALLOW
 import io.spine.logging.backend.given.FakeMetadata
-import io.spine.logging.jvm.given.FakeLogSite
+import io.spine.logging.given.FakeLogSite
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
