+++ logging/src/jvmTest/kotlin/io/spine/logging/SamplingRateLimiterSpec.kt
@@ -35,7 +35,7 @@ import io.spine.logging.LogContext.Key
 import io.spine.logging.RateLimitStatus.Companion.DISALLOW
 import io.spine.logging.backend.Metadata
 import io.spine.logging.backend.given.FakeMetadata
-import io.spine.logging.jvm.given.FakeLogSite
+import io.spine.logging.given.FakeLogSite
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
