+++ logging/src/jvmTest/kotlin/io/spine/logging/given/ConfigurableLogger.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023, The Flogger Authors; 2025, TeamDev. All rights reserved.
+ * Copyright 2025, TeamDev. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -24,14 +24,14 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.logging.jvm.given
+package io.spine.logging.given
 
 import io.spine.logging.Level
 import io.spine.logging.LogContext
 import io.spine.logging.LoggingApi
 import io.spine.logging.backend.LoggerBackend
 import io.spine.logging.AbstractLogger
-import io.spine.logging.jvm.given.ConfigurableLogger.Companion.DEFAULT_TIMESTAMP_NANOS
+import io.spine.logging.given.ConfigurableLogger.Companion.DEFAULT_TIMESTAMP_NANOS
 
 /**
  * Dynamically reconfigurable logger for testing backend
