+++ logging/src/jvmTest/kotlin/io/spine/logging/given/FakeLogSite.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.logging.jvm.given
+package io.spine.logging.given
 
 import io.spine.logging.LogSite
 import java.util.*
